python/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
distlib==0.3.8
filelock==3.13.3
idna==3.6
pipenv==2023.12.1
platformdirs==4.2.0
python-dotenv==1.0.1
requests==2.31.0
setuptools==69.2.0
urllib3==2.2.1
virtualenv==20.25.1
